refactor: update useListenBrowserUrlUpdates hook
