+++ packages/tuono/src/router/components/RouterProvider.tsx
@@ -1,6 +1,7 @@
 import { getRouterContext } from './RouterContext'
 import { Matches } from './Matches'
 import { useRouterStore } from '../hooks/useRouterStore'
+import { useListenBrowserUrlUpdates } from '../hooks/useListenBrowserUrlUpdates.tsx'
 import React, { useEffect, useLayoutEffect, type ReactNode } from 'react'
 
 type Router = any
@@ -71,34 +72,13 @@ const initRouterStore = (props?: ServerProps): void => {
   }, [])
 }
 
-const useListenUrlUpdates = (): void => {
-  const updateLocation = useRouterStore((st) => st.updateLocation)
-
-  const updateLocationOnPopStateChange = ({ target }: any): void => {
-    const { pathname, hash, href, search } = target.location
-    updateLocation({
-      pathname,
-      hash,
-      href,
-      searchStr: search,
-      search: new URLSearchParams(search),
-    })
-  }
-  useEffect(() => {
-    window.addEventListener('popstate', updateLocationOnPopStateChange)
-    return (): void => {
-      window.removeEventListener('popstate', updateLocationOnPopStateChange)
-    }
-  }, [])
-}
-
 export function RouterProvider({
   router,
   serverProps,
 }: RouterProviderProps): JSX.Element {
   initRouterStore(serverProps)
 
-  useListenUrlUpdates()
+  useListenBrowserUrlUpdates()
 
   return (
     <RouterContextProvider router={router}>
